refactor(dataset): extract restructure cluster from DatasetBag (Phase 3 §3.B)
Phase 3 audit §3.B — move the asset-restructuring cluster off
``DatasetBag`` into a focused module. The cluster takes a downloaded
bag and reorganizes its assets into a directory hierarchy suitable
for third-party ML trainers (e.g. ``torchvision.datasets.ImageFolder``).

WHY
====

``dataset_bag.py`` was a 2200-LoC file mixing two concerns:

  * Bag access — queries, member walking, parent/child traversal,
    feature reads, ML-framework adapters (as_torch_dataset, as_tf_dataset).
    ~1500 LoC, what users primarily interact with.
  * Asset restructuring — a single public function with ~500 LoC of
    body plus 4 private helpers, all in service of producing a
    directory hierarchy for ML trainers.

The restructuring cluster has one public entry (``restructure_assets``)
and reads from the bag's local SQLite via the helpers. Keeping it in
``dataset_bag.py`` made the bag class read like a laundry list.

WHAT MOVED
==========

New module ``src/deriva_ml/dataset/restructure.py`` (~750 LoC):

  * Module-level helper:
        _default_dir_name_from_target(target, targets)
            (was a free function in dataset_bag.py at line 67)
  * Free functions (each takes a DatasetBag arg; previously private
    methods on DatasetBag):
        _build_dataset_type_path_map(bag, type_selector)
        _get_asset_dataset_mapping(bag, asset_table)
        _get_reachable_assets(bag, asset_table)
        _detect_asset_table(bag)
  * Public free function (was DatasetBag.restructure_assets):
        restructure_assets(bag, output_dir, ...)

``DatasetBag.restructure_assets`` stays on ``DatasetBag`` as a thin
sugar method that delegates to the free function — preserves the
user-facing ``bag.restructure_assets(...)`` call shape.

DEAD-CODE SWEEP
================

Dropped ``DatasetBag._validate_dataset_types``. It was defined
(lines 1307-1332) but had no callers anywhere in src/ or tests/.
The audit's "no left over code from previous iterations" concern
was borne out here — looks like a leftover from a planning iteration
that was supposed to add validation to restructure_assets but never
landed.

WHY FREE FUNCTIONS, NOT A DatasetLike PROTOCOL
==============================================

The audit suggested lifting the cluster to take a ``DatasetLike``
rather than a concrete ``DatasetBag``. After investigation that
turned out to be aspirational:

  * _get_reachable_assets reads via bag.engine / bag._dataset_table_view
    — both bag-specific (the bag's local SQLite).
  * _build_dataset_type_path_map / _detect_asset_table COULD work on
    any DatasetLike (they use list_dataset_members, dataset_types,
    list_dataset_children, model — all on the protocol).
  * But cutting the cluster on the protocol-compatible vs.
    bag-specific line would leave half the cluster on DatasetBag
    and half in the module, with the module calling back into
    DatasetBag for the bag-specific bits. Worse shape than the
    full extraction.

Taking a concrete DatasetBag keeps the module symmetric with
``bag_download.py`` (which takes a Dataset) and avoids the half-split
awkwardness.

TEST MIGRATION
==============

``tests/test_restructure_assets.py`` was mocking the helpers as
methods on a ``MagicMock(spec=DatasetBag)``. After the extraction
the helpers are module-level functions, so the test was rewired to
monkey-patch the functions in ``deriva_ml.dataset.restructure``
directly:

  * An autouse fixture installs no-op stubs for the four helpers.
  * ``_make_bag`` rebinds ``_get_asset_dataset_mapping`` and
    ``_get_reachable_assets`` per-test so each test sees its own
    asset list.
  * Individual tests that needed custom helper returns (e.g.,
    Split-parent transparency) rebind directly on the module.

The MagicMock(spec=DatasetBag) is still used as a placeholder
"bag-like" object passed to the free function, but its method
attributes aren't accessed anymore.

NUMBERS
=======

  dataset_bag.py:    2170 LoC → 1507 LoC  (-663)
  restructure.py:    new      →   751 LoC

  Net: ~85 LoC (mostly the new module docstring + helper-function
  docstrings about the cluster, plus the sugar method's argument
  list duplicated as a forwarder).

TEST RESULTS
============

  * tests/test_restructure_assets.py: 11/11 passed
  * tests/local_db/ + tests/asset/ + tests/model/ +
    test_restructure_assets: 445 passed, 3 skipped, 0 failed
  * Lint (ruff): clean
